requirements.txt
selenium
rottentomatoes-python
bs4
pandas
scipy
seaborn
rapidfuzz